+++ lte/gateway/python/magma/monitord/icmp_state.py
@@ -18,7 +18,7 @@ from orc8r.protos.service303_pb2 import State
 ICMP_STATE_TYPE = "icmp_monitoring"
 
 ICMPMonitoringResponse = NamedTuple('ICMPMonitoringResponse',
-                                    [('last_reported_time', float),
+                                    [('last_reported_time', int),
                                      ('latency_ms', float)])
